refactor : put every component from expo\app\(tabs)\rooms\[id].tsx in separated files through bobby review
